fix: minor bugs

+++ src/v1/modules/influencer/models/index.ts
@@ -143,13 +143,13 @@ export class InfluencerProfile implements IInfluencerProfile {
   @Column({ type: 'varchar', default: null })
   location: string
 
-  @Column({ type: 'decimal', default: 0 })
+  @Column({ type: 'bigint', default: 0 })
   instagramCommercial: number
 
-  @Column({ type: 'decimal', default: 0 })
+  @Column({ type: 'bigint', default: 0 })
   twitterCommercial: number
 
-  @Column({ type: 'decimal', default: 0 })
+  @Column({ type: 'bigint', default: 0 })
   youtubeCommercial: number
 
   @CreateDateColumn({ type: 'timestamp' })
